% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shiny_metadata.R
\name{create_shiny_metadata}
\alias{create_shiny_metadata}
\title{Create metadata in Shiny}
\usage{
create_shiny_metadata(
  d = iris,
  var_criteria = list(short_description = "", foobar = Sys.Date())
)
}
\arguments{
\item{tba}{update_me.}
}
\description{
Create metadata in Shiny.
}
