\name{data_library}
\alias{data_library}
\title{Dataset Search}
\usage{
data_library()
}
\description{
Search for data set and display its summary value and scatter diagram in the desired field.
}
\examples{
data_library()
}


\name{help_data_library}
\alias{help_data_library}
\title{Dataset Search}
\usage{
help_data_library()
}
\description{
Search for help page of data sets in the desired field.
}
\examples{
help_data_library()
}
