% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{input}
\alias{input}
\title{Input data}
\usage{
input(
  path = getOption("data.library.path"),
  vars_regex = NULL,
  vars_vec = NULL,
  verbose = TRUE
)
}
\arguments{
\item{path}{Folder path of where the data will be output.}

\item{vars_regex}{Regex pattern to match variable names to input.}

\item{vars_vec}{A vector of of variable names to input.}

\item{verbose}{Logical argument to be verbose.}
}
\description{
Output data to *.RDS objects.
}
