% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shiny_metadata.R
\name{create_shiny_metadata}
\alias{create_shiny_metadata}
\title{Create metadata in Shiny}
\usage{
create_shiny_metadata(obj = letters, var_criteria)
}
\arguments{
\item{path}{Folder path of where the data will be output.}

\item{vars_regex}{Regex pattern to match variable names to input.}

\item{vars_vec}{A vector of of variable names to input.}

\item{verbose}{Logical argument to be verbose.}
}
\description{
Create metadata in Shiny.
}
